Upgrade slate js
Upgrade slate js to the alpha version 1. This makes things a lot easier to manage. Also added a Document outline at the start of an article. Created the dynamic article loader but have not implemented it yet because it is very slow. Added custom heading component and custom list item components